feat(repo): converted the repo into a turborepo
This change configurates next.js to run with turborepo. Second, it configures eslint to work with
jest and storybook, as well as adding good rules to keep code consistent across the repository. It
also adds some stylelint and postcss configuration.

BREAKING CHANGE: CI scripts will not work. CI and Vercel need to be configured to run with the new
code.